Force encode to ASCII when testing for end markers
The problem is caused here because vagrant wrapper is attempting to hunt
for a Vagrant binary by executing scripts on the system. These scripts
can return output with any encoding. Regardless of the string encoding
that comes off the system, using `%x` for execution the Ruby interpreter
**will** encode the result of the script as UTF-8.

The `#match` method assumes that it has been passed a String that is
valid in the passed encoding. This is, of course, highly incompatible
with systems that choose to output data in higher order bytes that
aren't actually unicode.

The solution presented in this patch simply invokes
`#force_encoding("ASCII-8BIT")` which will get rid of the potentially
invalid data. This is a bit of a hack, but it should work, because the
output marker being searched for is the string literal `"END VAGRANT
WRAPPER"`. It is very unlikely that any string encodings that Vagrant
users use will actually encode that ascii string differently as bytes.